Replaced the exon_set struct with a transcript struct
Previously I represented the sequences as a set of exon sequences
and their associated lengths (as a char** and int*). This has now
been replaced by representing the transcript as a single sequence
and keeping a record of the exon offsets and lengths instead. This
should make it much more straightforward to merge adjacent exons in
alignments.