Complete rewrite of alignment functions.
Complete re-implementation of the alignment functions. The sequence
of the transcript is now given as a single char*, with int arryas
indicating the lengths and offsets of the individual exons. Previously
I had represented the sequence as a set of exon char* pointers. The
new representation makes for a cleaner code, as concatenating exons
becomes a simple matter of modifying the exon length rather than trying
to obtain residues from different positions.