intron-constrained global alignments
Added a function that simply performs a Needleman-Wunsh that takes a
generalised substitution matrix and a specified special character.
Inserting this character at exon-exon borders allows and setting
appropriate substitution scores for this character results in an
alignment that will be biassed to conserving intron-exon boundaries.

The function also returns the positions of the special characters in the
alignments and a table giving the indices of aligned special characters.
These are the presumptive orthologous introns.