Added Smith-Waterman function
Added a Smith-Waterman function to needleman_wunsch.h and
needleman_wunsch.c. This compiles but has not been tested as
it still requires wrapper functions and function to extract
the alignment from the pointer matrix.

Obviously the names of files and the functions that they contain
no longer make any sense; I should probably split up the functions
into more files or rename the files.