Function to extract alignments from Smith Waterman tables.
Added a recursive function (extract_sw_alignments) that extracts
all non-overlapping alignments from the Smith-Waterman tables.

This function creates a quaternary tree of alignment objects; this
tree has to be converted to some sort of R structure, but I haven't
yet quite worked out how to do that.

I also added functions for clearing the memory used by such a tree and
for counting the number of entries.

None of these functions has yet been tested; there are likely to be
memory segmentation and other problems.