C functions for Smith Waterman alignments
Implementation of a Smith-Waterman alignment that reports all
non-overlapping alignments in the first sequence above a maximum
score.

The function compiles and seems to work reasonably well. It is not
rather slow taking about 1/100 second to extract alignments from a 660 x
337 nt alignment.

The function returns the complete scoring and pointer matrices along
with a whole load of other stuff.